Hysteresis feature
Let the user specify how many sequential detections
of "in range" or "out of range" occur before
executing a command.  Specifically, this should help
prevent the app from flipping into "out of range"
mode if the user's Bluetooth device happens to be
in a pocket or experiencing some other temporary
out of range condition.

Preferences Window: Allow user to specify number
of sequential detections of in or out of range before
executing the user specified command.

ProximityAppController - Set and load values for
in range and out of range detection count.

ProximityBluetoothMonitor - Count number of sequential
detections of the same state until user specified
threshold is cleared and then call foundDevice or
lostDevice.

README.md - Update version to 1.8