Now has 3 options for range + a way to get signal strength
For the range, I added a third option, Not Detectable.
So now:
0 = Out of Range, 1 = In Range, and 2 = Not Detectable

I changed isInRange to getRange because it fit better with the argument
I added, which is a way to get the RSSI to use in the Signal Bar, and
it still makes sense.

Also, I made the code a bit more unifirm with the rest :)